requirements.txt
rumps>=0.4.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
